requirements.txt
transformers
pillow
streamlit
torch
torchvision